As PTimedMutex::m_lockCount is always inside the mutex it is counting, it does not need to be atomic<>.

Tidied output of thread id's when PNullThreadIdentifier.
........
Fixed GNU compile
........
Add formatted output of JSON if set width in std:;ostream, which represents the indent level to start at.
........
Added protection against pthread_kill() crashing (insane!) when doing stack walk.
........
Fixed GNU compile
........

+++ src/ptlib/common/vconvert.cxx
@@ -1885,10 +1885,10 @@ bool PStandardColourConverter::YUV420PtoRGB(const BYTE * srcFrameBuffer,
 #endif
 
   BYTE * scanLinePtrRGB = dstFrameBuffer;
-  int scanLineSizeRGB = (rgbIncrement*m_dstFrameWidth+3)&~3;
+  int scanLineSizeRGB = (int)((rgbIncrement*m_dstFrameWidth+3)&~3);
 
   unsigned srcPixpos[4] = { 0, 1, m_srcFrameWidth, m_srcFrameWidth + 1 };
-  unsigned dstPixpos[4] = { 0, rgbIncrement, scanLineSizeRGB, scanLineSizeRGB+rgbIncrement };
+  unsigned dstPixpos[4] = { 0, rgbIncrement, (unsigned)scanLineSizeRGB, (unsigned)scanLineSizeRGB+rgbIncrement };
 
   if (m_verticalFlip) {
     scanLinePtrRGB += (m_dstFrameHeight - 2) * scanLineSizeRGB;
